Tiny signature fix

+++ src/readdata.jl
@@ -159,7 +159,7 @@ end
 let alphabet = [ 1,21, 2, 3, 4, 5, 6, 7, 8,21, 9,10,11,12,21,13,14,15,16,17,21,18,19,21,20]
                # A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y
     global letter2num
-    letter2num(c::UInt) = letter2num(Char(c))
+    letter2num(c) = letter2num(Char(c))
     function letter2num(c::Char)
         i = c - 'A' + 1
         1 <= i <= 25 && return alphabet[i]
